edina/crowdsurvey#16: First attempt at saving a completed Record.
Rather than saving in a separate index, the current implementation maps the array of Records that the Survey contains to the database.  

When a Survey is saved, a new Record object is created.  This takes the id and value params from the survey fields and creates new field instances to be associated with that Record instance.  This ensures that each Record has the correct field instances associated and isn’t just sharing the same Survey instances across all Records