edina/crowdsurvey#48: Improved record status logic.
Added functions to Record to to determine if all fields are empty and if all fields contain valid values.

Added vars to Field - containsValidValue and containsValue that are used by the Record to determine validity and set record state.

If a user does not complete any Fields in a form and goes back to the map, the empty Record that had been created will be discarded.