edina/crowdsurvey#53: Create RecordGeometry class for representing Record.geometry
This is mappable so can use ObjectMapper to ingest and serialise JSON and
used the CoordinateTransform class to facilitate this.

This class should be extended to support the MGLAnnotation protocol so
we can add annotations to the map view that are associated to a specific
record. This allows us to navigate back to record instances to edit them.