edina/crowdsurvey#56: Change location button depending on user location
When user selects the location button the button changes from outlined 
to filled (and the map pans to the user’s current location). Added pan 
gesture to MapView so that when the MapView is moved then the location 
icon is changed to outlined unless it already is outlined. 

Changed icon tint colour to white.